+++ admin/api/course.php
@@ -12,10 +12,21 @@ if (isset($_GET['call']))
     switch ($_GET['call']) 
     {
         case 'getAll':
+            if(isset($_GET['faculty_id']) && !empty($_GET['faculty_id'])) {
+                $faculty_id = $_GET['faculty_id'];
+                $response = $db->getAllByFaculty($faculty_id);
+                break;
+            }
             $response = $db->getAll();
             break;
 
         case 'add':
+            if(!isset($_POST['course_code']) || empty($_POST['course_code'])
+            || !isset($_POST['course_title']) || empty($_POST['course_title'])
+            || !isset($_POST['credit_hour']) || empty($_POST['credit_hour'])
+            || !isset($_POST['faculty_id']) || empty($_POST['faculty_id'])) {
+                break;
+            }
             $course_code = $_POST['course_code'];
             $course_title = $_POST['course_title'];
             $credit_hour = $_POST['credit_hour'];
@@ -28,6 +39,13 @@ if (isset($_GET['call']))
             break;
 
         case 'update':
+            if(!isset($_POST['id']) || empty($_POST['id'])
+            || !isset($_POST['course_code']) || empty($_POST['course_code'])
+            || !isset($_POST['course_title']) || empty($_POST['course_title'])
+            || !isset($_POST['credit_hour']) || empty($_POST['credit_hour'])
+            || !isset($_POST['faculty_id']) || empty($_POST['faculty_id'])) {
+                break;
+            }
             $id = $_POST['id'];
             $course_code = $_POST['course_code'];
             $course_title = $_POST['course_title'];
@@ -41,6 +59,9 @@ if (isset($_GET['call']))
             break;
 
         case 'delete':
+            if(!isset($_POST['id']) || empty($_POST['id'])) {
+                break;
+            }
             $id = $_POST['id'];
             $result = $db->delete($id);
 
@@ -50,6 +71,9 @@ if (isset($_GET['call']))
             break;
 
         case 'restore':
+            if(!isset($_POST['id']) || empty($_POST['id'])) {
+                break;
+            }
             $id = $_POST['id'];
             $result = $db->restore($id);
