+++ admin/api/donation.php
@@ -16,6 +16,12 @@ if (isset($_GET['call']))
             break;
 
         case 'add':
+            if(!isset($_POST['name']) || empty($_POST['name'])
+            || !isset($_POST['info']) || empty($_POST['info'])
+            || !isset($_POST['email']) || empty($_POST['email'])
+            || !isset($_POST['reference']) || empty($_POST['reference'])) {
+                break;
+            }
             $name = $_POST['name'];
             $info = $_POST['info'];
             $email = $_POST['email'];
@@ -28,6 +34,13 @@ if (isset($_GET['call']))
             break;
 
         case 'update':
+            if(!isset($_POST['id']) || empty($_POST['id'])
+            || !isset($_POST['name']) || empty($_POST['name'])
+            || !isset($_POST['info']) || empty($_POST['info'])
+            || !isset($_POST['email']) || empty($_POST['email'])
+            || !isset($_POST['reference']) || empty($_POST['reference'])) {
+                break;
+            }
             $id = $_POST['id'];
             $name = $_POST['name'];
             $info = $_POST['info'];
@@ -41,6 +54,9 @@ if (isset($_GET['call']))
             break;
 
         case 'confirm':
+            if(!isset($_POST['id']) || empty($_POST['id'])) {
+                break;
+            }
             $id = $_POST['id'];
             $result = $db->updateConfirmation($id, 1);
 
@@ -50,6 +66,9 @@ if (isset($_GET['call']))
             break;
 
         case 'unconfirm':
+            if(!isset($_POST['id']) || empty($_POST['id'])) {
+                break;
+            }
             $id = $_POST['id'];
             $result = $db->updateConfirmation($id, 0);
 
@@ -59,6 +78,9 @@ if (isset($_GET['call']))
             break;
 
         case 'delete':
+            if(!isset($_POST['id']) || empty($_POST['id'])) {
+                break;
+            }
             $id = $_POST['id'];
             $result = $db->delete($id);
 
@@ -68,6 +90,9 @@ if (isset($_GET['call']))
             break;
 
         case 'restore':
+            if(!isset($_POST['id']) || empty($_POST['id'])) {
+                break;
+            }
             $id = $_POST['id'];
             $result = $db->restore($id);
