+++ admin/api/student.php
@@ -27,6 +27,14 @@ if (isset($_GET['call']))
             break;
 
         case 'add':
+            if(!isset($_POST['name']) || empty($_POST['name'])
+            || !isset($_POST['id']) || empty($_POST['id'])
+            || !isset($_POST['reg']) || empty($_POST['reg'])
+            || !isset($_POST['batch_id']) || empty($_POST['batch_id'])
+            || !isset($_POST['session']) || empty($_POST['session'])
+            || !isset($_POST['faculty_id']) || empty($_POST['faculty_id'])) {
+                break;
+            }
             $name = $_POST['name'];
             $id = $_POST['id'];
             $reg = $_POST['reg'];
@@ -41,6 +49,14 @@ if (isset($_GET['call']))
             break;
 
         case 'update':
+            if(!isset($_POST['name']) || empty($_POST['name'])
+            || !isset($_POST['id']) || empty($_POST['id'])
+            || !isset($_POST['reg']) || empty($_POST['reg'])
+            || !isset($_POST['batch_id']) || empty($_POST['batch_id'])
+            || !isset($_POST['session']) || empty($_POST['session'])
+            || !isset($_POST['faculty_id']) || empty($_POST['faculty_id'])) {
+                break;
+            }
             $name = $_POST['name'];
             $id = $_POST['id'];
             $reg = $_POST['reg'];
@@ -55,6 +71,9 @@ if (isset($_GET['call']))
             break;
 
         case 'delete':
+            if(!isset($_POST['id']) || empty($_POST['id'])) {
+                break;
+            }
             $id = $_POST['id'];
             $result = $db->delete($id);
 
@@ -64,6 +83,9 @@ if (isset($_GET['call']))
             break;
 
         case 'restore':
+            if(!isset($_POST['id']) || empty($_POST['id'])) {
+                break;
+            }
             $id = $_POST['id'];
             $result = $db->restore($id);
